--- app/assets/javascripts/active_admin.js.coffee
@@ -1,1 +1,0 @@
-#= require active_admin/base
