+++ app/views/categories/edit.html.slim
@@ -1,12 +1,1 @@
-.container
-  .col
-    .card-body.bg-light 
-        h1 Edit category
-        form
-        = form_for :category, url: category_path(@category), method: :patch do |f|
-            .form-group
-                = f.label :name,'Name'
-                br
-                = f.text_field :name, class: 'form-control', placeholder: 'Name'
-            br
-            = f.submit "Save Category", :style =>"width: 400px", class: 'btn btn-lg btn-primary btn-block'
+render = render partial: "form"
