+++ app/views/products/_form.html.slim
@@ -1,0 +1,56 @@
+.card-body.bg-light
+    -if @product
+        h1 Editing a product
+        form
+        = form_for @product do |f| 
+            .form-group
+                = f.label :name,'Name'
+                br
+                = f.text_field :name, class: 'form-control', placeholder: 'Name'
+            .form-group
+                = f.label :shop,'Shop'
+                br
+                = f.text_field :shop, class: 'form-control', placeholder: 'Shop'
+            .form-group
+                = f.label :category,'Category'
+                br
+                = select_tag :category, options_for_select(@choices, @choice), class: 'form-control'
+            .form-group
+                = f.label :price,'Price'
+                br
+                = f.text_field :price, class: 'form-control', placeholder: 'Price'
+            .form-group
+                = f.label :quantity,'Quantity'
+                br
+                = f.text_field :quantity, class: 'form-control', placeholder: 'Quantity'
+            br
+            = f.submit "Save Product", :style =>"width: 300px", class: 'btn btn-lg btn-primary btn-block'
+           
+    -else
+        h1 Adding a product
+        form
+        = form_for :product, url: products_path do |f| 
+            .form-group
+                = f.label :name,'Name'
+                br
+                = f.text_field :name, class: 'form-control', placeholder: 'Name'
+            .form-group
+                = f.label :shop,'Shop'
+                br
+                = f.text_field :shop, class: 'form-control', placeholder: 'Shop'
+            .form-group
+                = f.label :category,'Category'
+                br
+                = select_tag :category, options_for_select(@choices, @choice), class: 'form-control'
+            .form-group
+                = f.label :price,'Price'
+                br
+                = f.text_field :price, class: 'form-control', placeholder: 'Price'
+            .form-group
+                = f.label :quantity,'Quantity'
+                br
+                = f.text_field :quantity, class: 'form-control', placeholder: 'Quantity'
+            br
+            = f.submit "Save Product", :style =>"width: 300px", class: 'btn btn-lg btn-primary btn-block'
+            br
+            = link_to "+Add category", categories_path, :style =>"width: 300px", class: 'btn btn-lg btn-secondary btn-block'
