+++ app/views/products/edit.html.slim
@@ -1,27 +1,2 @@
 
-.card-body.bg-light 
-    h1 Edit a product
-    form
-    = form_for :product, url: product_path(@product), method: :patch do |f|
-        .form-group
-            = f.label :name,'Name'
-            br
-            = f.text_field :name, class: 'form-control', placeholder: 'Name'
-        .form-group
-            = f.label :shop,'Shop'
-            br
-            = f.text_field :shop, class: 'form-control', placeholder: 'Shop'
-        form-group
-            = f.label :category,'Category'
-            br
-            = f.text_field :category, class: 'form-control', placeholder: 'Category'
-        form-group
-            = f.label :price,'Price'
-            br
-            = f.text_field :price, class: 'form-control', placeholder: 'Price'
-        form-group
-            = f.label :quantity,'Quantity'
-            br
-            = f.text_field :quantity, class: 'form-control', placeholder: 'Quantity'
-        br
-        = f.submit "Save Product", :style =>"width: 400px", class: 'btn btn-lg btn-primary btn-block'
+ render = render partial: "form"
