+++ db/schema.rb
@@ -68,7 +68,6 @@ ActiveRecord::Schema.define(version: 2018_11_08_211937) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "username"
-    t.boolean "admin", default: false
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
